Step 2 - Example
Create an example interface that defines the properties of the value
object to be generated. This example will serve as a test for the code
template to be written. Without such a test, we won't get any error
messages if something is wrong in with the template.
The example is taken from
https://github.com/google/FreeBuilder#quick-start.