Step 5 - Create a private field for each property
To generate one ore multiple fields, you need to add a field to the
template and annotate it with @Field.

#{src.properties} is a Java EL expression that means "all properties of
class IPerson", since IPerson was the "src" we started with. 

IPerson is a javax.lang.model.element.TypeElement.
So, in the Java EL expression you have access to all properties of
TypeElement. TypeElement does not have any property with the name
"properties". However, "properties" is some convenient extension that
japkit provides for TypeElements in EL expressions. It is a list of
de.japkit.model.Property.

Since #{src.properties} is a collection, a field will be generated for
every element in the collection, that is for every property of class
IPerson.
While japkit iterates through that collection, it will set the "src"
variable to each element of the collection. So, the new "src" for
generating each field is a de.japkit.model.Property.

The field is private and it shall have the type and name of the "src"
element (the property):

- SrcType is a type function and means to use the type of the "src"
element (in Java EL, it would be "#{src.asType()}")
- The $...$ syntax tells japkit to insert the result of an expression
evaluation here. $name$ is short for $simpleName$, and since "src" is
implicit in expressions, this is the same as #{src.simpleName} . So the
name of the property is inserted here.

Take a look at the generated class Person:
On each generated field, there is an @Order annotation.
These @Order annotations provide a deterministic order for the members
of the generated class when they are used by subsequent generators. This
is required during Eclipse incremental compilation or if a code template
in another maven project accesses the class. In both cases, only binary
type information might be provided to the annotation processor, and
member order would not be well-defined.