Step 6 - Generate getters and private setters
Generate getters and private setters for the field. The setters will be
used later by the generated builder class.
You can generate arbitrary methods with japkit, but getters and setters
are so common, that there are convenient @Getter and @Setter annotations
to generate the accessor methods for a field.

We also tell japkit to generate an implements clause here for the
SrcIterface which is IPerson in our example.

In the generated class, the @ParamNames annotations have a similar
purpose as @Order. They provide method parameter names for cases where
only binary type information is provided to the annotation processor.
You can simply ignore it.