Step 8 - Fields, getters and getters for Builder
Generate fields, getters and setters for the Builder class. Pretty much
the same as in the value object, but the setters are public and can be
chained (i.e. they return this). Note that the setter still use Java
Beans naming convention. So, the Builder can be used with data binding
frameworks (Jackson, Spring MVC, ...).