Step 9 - The build method
At first, remember the type of the generates value object class
(Person). Then, generate the build method. It's code body creates an
instance of the value object and then sets each property of the value
object by calling the private setters. Finally, the value object is
returned. 
Note, that "src" can be omitted in most places, for example "properties"
instead of "src.properties".
The "code" in "valueObjectClass.code" takes care for proper
representation of the qualified type name in the generated code. If
possible, an import statement is added.