Step 11.2 - Immutability: Matcher functions
To decide which code to generate for defensive copies we have to
distinguish between the relevant property types (here Date, List and
Set). For this, we define some Matcher functions. A Matcher takes the
current source element as input and checks for the given criteria. If it
matches, true is returned, otherwise false. Here, we check for the type
of the element. There are several other possible criteria, for example
checking for existence of an annotation on the element. 
Note that Matchers are only a special case of functions. You can use the
@Function annotation to define any other kinds of functions.
We define the functions isDate, isList and isSet as annotations. This
eases use of them in switch statements as we will see in next steps.
The functions could be defined in ValueObjectTemplate, but since they
might be of general use when generating domain objects, we put them into
a separate class DomainLibrary.