Step 11.3 - Immutability: Getters
The getters of our value object must not return mutable data structures
(here: Date, List and Set) that allow to change the state of the value
object. Thus we must create defensive copies of the field value and / or
make them unmodifiable.
At first, we extend our example IPerson to have properties of type Date
and List.
Then, we remove the default getter from the template an define our own.
Its code body uses the Switch function 'getterRhs', which generates the
appropriate code for each type of property.
We use the Matcher functions isDate, isList and isSet defined earlier as
conditions for the Switch. The Switch functions allows to use any
boolean function as condition and it supports to refer to them by
annotations. For more complex cases, the @Case annotation can be used
instead.
CodeFragment is another special function. It returns a String, but it
additionally supports import handling for type names. Here, we import
our ValueObjectUtils and the java.util.Collections class.
The code bodies we have seen so far for the build method and the
constructors are CodeFragments as well, but for convenience, they are
"integrated" in the @Method and @Constructor annotations.
Since all the cases of the Switch are CodeFragments here, the Switch
itself becomes also a code fragment which makes it a proper candidate to
be used in the code body of the getter.
The DefaultCase is applied if the other conditions are false. In this
case, we just return the field without making a defensive copy.
Note that the field types (String) have no special meaning here.
Basically, the fields in the Switch are just the place to put the
annotations.