Step 11.4 - Immutability: Setters
The private setters of our value object need to make defensive copies of
the values they get from the builder. Otherwise, the value object could
be initialized with Lists, Sets or Dates that are modified afterwards by
the code that created them.
The code template changes are similar to the previous step. This time,
we use the immutableCopy methods implemented earlier in
ValueObjectUtils.