Step 11.5 - Refactoring: Template for properties
Meanwhile, we iterate over "src.properties" in three places: the field,
the getter and the setter. Thus we define a Template which contains
those elements and has "src.properties" as source.
Note that Template will NOT generate an inner class but it is just a way
to encapsulate multiple elements of a code template.
The Template must not be static, since static templates are not applied
immediately, but must be explicitly applied with @TemplateCall.
Templates can be nested, if necessary.