Step 12 - Generate equals and hashCode
This is pretty much the same approach as for the build method: We
iterate over the properties an generate some code for each of them.
However, for equals it would be quite cumbersome to put the
bodyBeforeIteratorCode in an annotation value, since it is multi-lined.
Thus, japkit allows to write the values for bodyBeforeIteratorCode,
bodyCode, bodyAfterIteratorCode into the Javadoc of the method template.
You have to use the ul / li syntax as shown here and if you have mutiple
lines to generate they must be wrapped by pre. Japkit regognizes this
special syntax and removes it from the Javadoc before copying it to the
generated method.
The main reason for the ul / li / pre syntax is to prevent unwanted
reformatting by the IDE.
We also added the @RuntimeMetadata annotation to ValueObjectTemplate. By
this, the values from the Javadoc are also available, if japkit has no
access to the source code of ValueObjectTemplate.
This happens for example, if you put ValueObjectTemplate into another
Maven project. But also during incremental compilation in Eclipse there
are situations where only binary type information of ValueObjectTemplate
(i.e. the .class file) is available for japkit.