Step 13 - Generate toString
For the toString method, we again use the Javadoc syntax, since we had
to escape the quotes if we used annotation values, which would make the
template harder to read.
As alternative to the li / ul pre syntax, the Javadoc tag syntax is used
here. But note that it is less resistant to code formatting in Eclipse,
especially when using pre for multi-lined code.