Step 14.2 - JPA annotations: Optionally use @Entity
JPA Embeddables have some limitations. For example, an embeddable class
contained within an element collection must not contain an element
collection.
Thus, it is sometimes necessary to implement a value object as a JPA
Entity instead.
We want to make this configurable, so we add the asJpaEntity annotation
value to @ValueObject.
The annotation values of the trigger annotation are made available as
variables to be used in Java EL expression. So, we can use asJpaEntity
to conditionally switch between @Embeddable and @Entity and to generate
the @Id only if it is an Entity.
The @Embeddable_ and @Entity_ are so called annotation templates. They
have all the fields of the original annotations, but additionally they
have fields like "_cond" which allows to generate the annotation
conditionally.
The annotation templates for JPA come out of the box with japkit, but
they can be easily created for other annotations as well. See for
example https://github.com/stefanocke/japkit/blob/master/japkit-annotationtemplates/src/main/java/de/japkit/annotationtemplates/JpaAnnotationTemplates.java