Step 14.3 - JPA annotations: More Matcher functions
As preparation for the next steps, we introduce a bunch of new Matcher
functions to distinguish between several types of fields in a value
object. 
There are some new concepts here:
- Boolean functions (like Matchers) can be combined by @And, @Or, @Not.
- japkit provides a convenient property "singleValueType" to determine
the element type of a collection (like Set or List). Similar, Matchers
can operate on the singleValueType when matching for annotations
("singleValueTypeAnnotations"). 
- "asElement().ValueObject" shows how to easily access an elements
annotation by its simple name (ValueObject). To make this work, you have
to add a so called annotation import to the library.