Step 14.4 - JPA annotations: @Embedded and @ElementCollection
We add the @Embedded annotation to a value object's field, of the type
of the field is an Embeddable and if is not a collection. To test this,
we create the new value object Address an add it to Person. 
We add the @ElementCollection annotation to a value object's field, if
is collection typed and the collections element type is a JPA basic type
or an Embeddable. So, our telephoneNumbers field, which is a
List<String>, becomes an ElementCollection.
For fields of type List, we also add an @OrderColumn to preserve the
list order.