Step 14.5 - JPA annotations: @Temporal and @Enumerated
We also add JPA annotations for fields with temporal type (like birthday
in IPerson) and enumeration types.
While it might be debatable, if all temporal fields should be mapped as
timestamps in the database, it is certainly a good idea to have a common
strategy how to map enums (as STRING or ORDINAL, but not both) within
the whole project.
If necessary, the mapping for temporal types could be fine tuned by
using an additional annotation in IPerson. But in general, the code
template should only give as much freedom as really required and enforce
as much conventions as possible.