Step 14.6 - JPA annotations: @OneToOne and @OneToMany
When the type of a property is another value object, but this value
object is implemented as JPA entity (see Step 14.1), we cannot put
@Embeddable or @ElementCollection on the field but must use @OneToOne
and @OneToMany.
Additionally, we set CascadeType.ALL and orphanRemoval, since we want to
have a real containment relationship here: If a value object is deleted,
all its constituents shall also be deleted. Changes shall also be
propagated.

Note that we have also added our DomainLibrary in the @Clazz annotation.
Primarily this is required to use the functions of the library in Java
EL expressions. But it is currently also required to make the annotation
import for isValueObjectAsJpaEntity work (see step 14.3).