Step 14.7 - JPA annotations: @ManyToOne and @ManyToMany
When the type of a property is a JPA entity, but not a value object, we
put @ManyToOne and @ManyToMany on the field.
Note that it can never be a @OneTo...  relationship here, since the
other Entity cannot "belong" to the value object.
These kind of restrictions do not come from JPA. They are "coded
architecture" originating from the decision to implement entities, value
objects and aggregates in the sense of Domain Driven Design.

To test the new annotations, we create the City entity and use it in the
Address value object.